pyproject.toml
[project]
name = "translateprocessual"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "extra-streamlit-components>=0.1.81",
    "google-genai>=1.56.0",
    "google-generativeai>=0.8.6",
    "numpy>=2.4.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.2",
    "streamlit-cookies-manager>=0.2.0",
    "supabase>=2.27.1",
]

[theme]
base="light"
primaryColor="#2723ed"
backgroundColor="#FFFFFF"
secondaryBackgroundColor="#F8FAFC"
textColor="#0F172A"
font="sans serif"